pyproject.toml
[project]
name = "symmetrical-train"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "datasets>=4.4.1",
    "fastapi[standard]>=0.121.2",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.38.0",
    "unidecode>=1.3.8",
]
